pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cts"
version = "0.1.0"
description = "Configurable capability transit system for APIs, MCP, CLI tools, and scripts."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "click>=8.1,<9",
  "httpx>=0.27,<1",
  "jsonschema>=4.22,<5",
  "pydantic>=2.7,<3",
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]

[project.scripts]
cts = "cts.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

